checks: be less aggressive in version checks; consolidate logic
